Refactors calendar logic into service
